Rename 'picobox.contrib' to 'picobox.ext'
I realized that I used the 'contrib' subpackage improperly. Let me quote
Drew DeVault on what 'contrib' actually is:

> The contrib directory is used as an unorganized (or, at best, lightly
> organized) bin of various useful things contributed by the community
> around the software, but which is not necessarily a good candidate for
> being a proper part of the software. hings in contrib should not be
> wired into your build system, shouldn’t be part of your automated
> testing, shouldn’t be included in your documentation, and should not be
> installed with your packages. contrib entries are not supported by the
> maintainers, and are given only a light code review at the most.
>
> https://drewdevault.com/2020/06/06/Add-a-contrib-directory.html

This patch renames the `picobox.contrib` subpackage into `picobox.ext`,
and this subpackage is intended to contain various extensions to the
picobox core functionality, similar to how `sqlalchemy.ext` contains
various extensions for SQLAlchemy core.
